#192 Extract InMemoryCacheAction to a separate class
