+++ action/library/src/test/java/io/knotx/fragments/action/library/logging/InMemoryCacheActionFactoryLoggingTest.java
@@ -17,13 +17,12 @@
  */
 package io.knotx.fragments.action.library.logging;
 
-
-import static io.knotx.fragments.action.library.InMemoryCacheActionFactory.CACHED_VALUE;
-import static io.knotx.fragments.action.library.InMemoryCacheActionFactory.CACHE_HIT;
-import static io.knotx.fragments.action.library.InMemoryCacheActionFactory.CACHE_KEY;
-import static io.knotx.fragments.action.library.InMemoryCacheActionFactory.CACHE_MISS;
-import static io.knotx.fragments.action.library.InMemoryCacheActionFactory.CACHE_PASS;
-import static io.knotx.fragments.action.library.InMemoryCacheActionFactory.COMPUTED_VALUE;
+import static io.knotx.fragments.action.library.InMemoryCacheAction.CACHED_VALUE;
+import static io.knotx.fragments.action.library.InMemoryCacheAction.CACHE_HIT;
+import static io.knotx.fragments.action.library.InMemoryCacheAction.CACHE_KEY;
+import static io.knotx.fragments.action.library.InMemoryCacheAction.CACHE_MISS;
+import static io.knotx.fragments.action.library.InMemoryCacheAction.CACHE_PASS;
+import static io.knotx.fragments.action.library.InMemoryCacheAction.COMPUTED_VALUE;
 import static io.knotx.fragments.api.FragmentResult.ERROR_TRANSITION;
 import static io.knotx.fragments.api.FragmentResult.SUCCESS_TRANSITION;
 import static io.knotx.junit5.assertions.KnotxAssertions.assertJsonEquals;
